fix build, no longer store file locations for qt translations

+++ Source/relive_lib/Camera.hpp
@@ -21,6 +21,6 @@ public:
     s16 mCamYOff = 0;
     s16 mPath = 0;
     EReliveLevelIds mLevel;
-    s16 mCameraNumber = 0;
+    u32 mCameraNumber = 0;
     bool mCamResLoaded = false;
 };
